[R] Modified QMake file SPH.pro to repair MinGW configurations,       which broke after switching to MSVC and changing QT version. [M] Cleansed code and made it compatible with both MSVC and MinGW       building. Achieved that by using C preprocessor directives       (macros). [+] Added some new "switches" in Settings.cpp. This includes       "NO_PRINTOUT" and "NO_SCREENS", the second of which eliminates       inconsistent framerate when enabled (framerate drops       inconsistently when every frame is being saved as a file,       which has been the default and only behaviour, so far).
[R] - repair / correction,
[M] - modification / refactoring,
[+] - new feature,
[-] - removal / revert.

+++ util/strings.cpp
@@ -1,3 +1,4 @@
+#include "util/settings.h"
 #include "util/strings.h"
 
 #include <QDir>
@@ -10,11 +11,14 @@ QString Strings::DIR_FRAMES;
 
 void Strings::init()
 {
-    QDir(Strings::DESKTOP + QString("/frames/")).removeRecursively();
-    QDir().mkdir(Strings::DESKTOP + QString("/frames/"));
-    DIR_FRAMES
-        = QStandardPaths::locate(QStandardPaths::DesktopLocation, "frames/",
-                                 QStandardPaths::LocateDirectory);
+    if (! Settings::NO_SCREENS)
+    {
+        QDir(Strings::DESKTOP + QString("/frames/")).removeRecursively();
+        QDir().mkdir(Strings::DESKTOP + QString("/frames/"));
+        DIR_FRAMES
+            = QStandardPaths::locate(QStandardPaths::DesktopLocation, "frames/",
+                                     QStandardPaths::LocateDirectory);
+    }
 }
 
 Strings::Strings()
